requirements.txt
azure-functions
fastapi==0.111.0
uvicorn==0.29.0
SQLAlchemy==2.0.25
PyJWT==2.8.0
PyMySQL==1.1.0